+++ src/app/shared/shared.module.ts
@@ -5,7 +5,8 @@ import { RouterModule } from '@angular/router';
 import { RoundButton } from 'src/app/shared/components/round-button/fs-round-button.component';
 import { OfferTileComponent } from 'src/app/shared/components/offer-tile/offer-tile.component';
 import { ErrorBackTextComponent} from 'src/app/shared/components/error-back-text/error-back-text.component';
-import { LoadingSpinnerComponent } from 'src/app/shared//components/loading-spinner/loading-spinner.component';
+import { LoadingSpinnerComponent } from 'src/app/shared/components/loading-spinner/loading-spinner.component';
+import { InfoBannerComponent } from 'src/app/shared/components/info-banner/info-banner.component';
 
 import {
   FaIconLibrary,
@@ -27,13 +28,13 @@ import {
   faCog
 } from '@fortawesome/free-solid-svg-icons';
 
-
 @NgModule({
 declarations: [
   RoundButton,
   OfferTileComponent,
   ErrorBackTextComponent,
-  LoadingSpinnerComponent
+  LoadingSpinnerComponent,
+  InfoBannerComponent
 ],
 imports: [
   CommonModule,
@@ -48,7 +49,8 @@ exports: [
   RoundButton,
   OfferTileComponent,
   ErrorBackTextComponent,
-  LoadingSpinnerComponent
+  LoadingSpinnerComponent,
+  InfoBannerComponent
 ]
 })